Architecture updates funtional. Updated routing to enable a multi-page website
