+++ package.json
@@ -9,17 +9,16 @@
     "@testing-library/user-event": "^13.5.0",
     "animate.css": "^4.1.1",
     "bootstrap": "^5.2.3",
-    "dotenv": "^16.0.3",
     "express": "^4.18.2",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.0",
     "react-bootstrap-icons": "^1.10.2",
     "react-dom": "^18.2.0",
     "react-icons": "^4.7.1",
-    "react-multi-carousel": "^2.8.2",
     "react-on-screen": "^2.1.1",
+    "react-router-dom": "^6.20.0",
     "react-scripts": "5.0.1",
-    "sass": "^1.57.0",
+    "styled-components": "^6.1.1",
     "web-vitals": "^2.1.4"
   },
   "scripts": {
