+++ src/components/Banner.js
@@ -47,7 +47,7 @@ export const Banner = () => {
   }
 
   return (
-    <section className="banner" id="home">
+    <section className="banner">
       <Container>
         <Row className="align-items-center">
           <Col xs={12} md={12} xl={12}>
