+++ src/components/NavBar.js
@@ -2,14 +2,14 @@ import { useEffect, useState } from 'react';
 import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
+import {Link} from 'react-router-dom';
 
 import logo from '../assets/img/logo2.svg';
 import navIcon1 from '../assets/img/nav-icon1.svg';
 import navIcon2 from '../assets/img/nav-icon2.webp';
-import navIcon3 from '../assets/img/nav-icon3.svg';
 
 export const NavBar = () => {
-    const [activeLink, setActiveLink] = useState('home');
+
     const [scrolled, setScrolled] = useState(false);
 
     useEffect(() => {
@@ -25,32 +25,26 @@ export const NavBar = () => {
         return () => window.removeEventListener('scroll', onScroll);
     }, [])
 
-    const onUpdateActiveLink = (value) => {
-        setActiveLink(value);
-    }
-
     return (
         <Navbar expand="lg" className={scrolled ? 'scrolled' : ''}>
             <Container>
                 <Navbar.Brand href="#home">
                     <img src={logo} alt='logo' />
                 </Navbar.Brand>
-                <Navbar.Toggle aria-controls="basic-navbar-nav">
-                    <span className='navbar-toggler-icon'></span>
-                </Navbar.Toggle>
-                <Navbar.Collapse id="basic-navbar-nav">
-                <Nav className="me-auto">
-                    <Nav.Link href="#home" className={activeLink === 'home' ? 'active navbar-link' : 'navbar-link'} onClick={() => onUpdateActiveLink('home')}>Home</Nav.Link>
-                    <Nav.Link href="#skills" className={activeLink === 'skills' ? 'active navbar-link' : 'navbar-link'} onClick={() => onUpdateActiveLink('skills')}>Skills</Nav.Link>
-                    <Nav.Link href="#projects" className={activeLink === 'projects' ? 'active navbar-link' : 'navbar-link'} onClick={() => onUpdateActiveLink('projects')}>Projects</Nav.Link>
+
+                <Navbar.Collapse>
+                <Nav>
+                    <Link to="/">Home</Link>
+
+                    <Link to='/support'>Support</Link>
+
+                    <Link to='/contact'>Contact</Link>
                 </Nav>
                 <span className='navbar-text'>
                     <div className='social-icon'>
                         <a href='https://www.linkedin.com/in/nicolette-renner/'><img src={navIcon1} alt="LinkedIn Icon" /></a>
                         <a href='https://github.com/nrenner0211'><img src={navIcon2} alt="GitHub Icon" /></a>
-                        <a href='https://www.facebook.com/niiiiiiiiiiiiiiiiiiiiiiiik/'><img src={navIcon3} alt="Facebook Icon" /></a>
                     </div>
-                    {/* <button><a href="https://www.linkedin.com/in/nicolette-renner/"><span>Let's Connect</span></a></button> */}
                 </span>
                 </Navbar.Collapse>
             </Container>
