--- src/components/Skills.js
@@ -1,75 +1,0 @@
-import meter1 from "../assets/img/meter1.svg";
-import meter2 from "../assets/img/meter2.svg";
-import meter3 from "../assets/img/meter3.svg";
-import Carousel from 'react-multi-carousel';
-import 'react-multi-carousel/lib/styles.css';
-
-export const Skills = () => {
-  const responsive = {
-    superLargeDesktop: {
-      breakpoint: { max: 4000, min: 3000 },
-      items: 5
-    },
-    desktop: {
-      breakpoint: { max: 3000, min: 1024 },
-      items: 3
-    },
-    tablet: {
-      breakpoint: { max: 1024, min: 464 },
-      items: 2
-    },
-    mobile: {
-      breakpoint: { max: 464, min: 0 },
-      items: 1
-    }
-  };
-
-  return (
-    <section className="skill" id="skills">
-        <div className="container">
-            <div className="row">
-                <div className="col-12">
-                    <div className="skill-bx wow zoomIn">
-                        <h2>Skills</h2>
-                        <p>Professionally trained to design accessible, mobile-first applications and to produce excellent documentation. Here are a few of my most used skills and tools:</p>
-                        <Carousel responsive={responsive} infinite={true} className="owl-carousel owl-theme skill-slider">
-                            <div className="item">
-                                <img src={meter1} alt="" />
-                                <h5>React & Next.js</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter1} alt="" />
-                                <h5>Node.js</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter1} alt="" />
-                                <h5>Auditing with Lighthouse</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter3} alt="" />
-                                <h5>Web Accessibilty Standards</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter3} alt="" />
-                                <h5>API Integration</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter3} alt="" />
-                                <h5>MongoDB & MySQL</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter2} alt="" />
-                                <h5>Jest, Postman, Selenium Testing</h5>
-                            </div>
-                            <div className="item">
-                                <img src={meter2} alt="" />
-                                <h5>Logo Design</h5>
-                            </div>
-                        </Carousel>
-                    </div>
-                </div>
-            </div>
-        </div>
-    </section>
-  )
-}
